Moved routines from SSE namespace to SSE2 namespace
